Refactor TreeNode.text: string to TreeNode.depth: number
Because TreeNode.text can contain only "root", "child", "grandchild",
"grandgrandchild", ..., it can be replaced with integers.

The logic to convert a depth to a text is moved from handleAddChild to a
new function nodeToDepthName.

nodeToLabel was renamed to nodeToDescendantInfo to clarify its intent.

nodeIdToColorDict was replaced with a function depthToColor.